Adds new configuration class to separate Balance Cache initialization from the controller
